Delete the unused MultiTap
